checkout-flask/requirements.txt
flask>=3.0
vonpay-checkout>=0.4.0
